MTH229Lite

Documentation for MTH229Lite.

MTH229Lite.annotate!

MTH229Lite.plot

MTH229Lite.plot!

MTH229Lite.plotif

MTH229Lite.scatter

MTH229Lite.title!

MTH229Lite.annotate! — Method

annotate!([p::Plot], x, y, txt; size, color, textposition, kwargs...)
annotate!([p::Plot], anns::Tuple; size, color, textposition, kwargs...)

Add annotations to plot

size: text size

color: text color

textposition: one of "top", "bottom", "left", "right", or combinations with _

MTH229Lite.plot! — Method

plot!([p::Plot], f; kwargs...)
plot!([p::Plot], x, y; kwargs...)

Used to add a new tract to an existing plot. Like Plots.plot!

MTH229Lite.plot — Method

plot(x, y; kwargs...)
plot(f::Function, a, b; kwargs...)

Create a plot. Returns a Plot instance from PlotlyLight

Provides an interface like Plots.plot for plotting a function f using PlotlyLight. This just scratches the surface, but PlotLight allows full acces to the underlying JavaScript library.

The provided "Plots" like functions are plot, plot!, scatter, scatter!, annotate, annotate!, title!.

Example

plot(sin, 0, 2pi)
plot!(cos)
x0 = [pi/4, 5pi/4]
scatter!(x0, sin.(x0), markersize=10)
title!("Sine and cosine and where they intersect in [0,2π]")

MTH229Lite.plotif — Method

plotif(f, g, a, b)

Plot f colored depending on g >= 0 or not.

MTH229Lite.scatter — Method

scatter(x, y; markershape, markercolor, markersize, kwargs...)
scatter!([p::Plot], x, y; kwargs...)

Place point on a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red"

markersize: size, as an integer

MTH229Lite.title! — Method

title!([p::Plot], txt)

Set plot title.