MTH229Lite.MTH229Lite — Module

MTH229Lite

Support files for MTH229. Unlike the MTH229 package, this package omits SymPy and has a simplified plotting setup for running within resource-constrained environments, such as binder.org (e.g. mth229lite).

Exports:

e is aliased to exp(1)

rangeclamp to restrict large y values; fisheye to change x axis scale; unzip to manipulate plotting limits

lim to explore limits numerically

sign_chart to explore sign changes numerically

bisection and newton for solving f(x) = 0. This package also exports Roots which exposes find_zero and find_zeros (and their aliases fzero and fzeros).

tangent(f,c) and secant(f,a,b) for creating tangent and secant line functions

riemann a simple Riemann sum function. This package also exports QuadGK for real-world usage.

The operator ' is overloaded for functions to find the derivative using the ForwardDiff package. See D also.

The PlotlyLightLite package provides a basic plotting interface to PlotlyLight.

The SimpleExpressions package is included to create simple expressions to use in place of functions. There is no SymPy package provided, as is done with MTH229.

MTH229Lite.Interval — Type

Interval(a,b)

A means to specify an interval in the form a..b

There are methods for

plot(eq, I) (two plots)

solve(eq, I) (dispatch to find_zeros)

sign_chart, riemann, quadgk

There are parsing gotchas

-1..2 is okay

-2..-1 is not, try -2..(-1)

-1..x^2 will parse as Interval(-1,x^2), also -1..x+2 becomes Interval(-1, x+2), as .. has low precedence.

Displays with unicode brackets to disambiguate from vector or tuple:

julia> -2..1
⟦-2, 1⟧

CommonSolve.solve — Method

solve(ex::SymbolicEquation, x₀, args...; kwargs...)
solve(ex::SymbolicEquation, x₀::Interval; kwargs...)`

Numerically solve an equation specified with a @symbolic value.

The find_zero or find_zeros function is used, the latter when x₀ is of type Interval (specified with the .. infix operator.

Example

julia> solve(e^x ~ x^4, 10)       # solution near 10
8.6131694564414

julia> solve(e^x ~ x^4, -10, 10)  # some solution in bracketing interval
-0.8155534188089607

julia> solve(e^x ~ x^4, -10..10)  # all solutions within [-10, 10]
3-element Vector{Float64}:
 -0.8155534188089606
  1.4296118247255554
  8.6131694564414

CommonSolve.solve — Method

solve(ex::SymbolicEquation)

Tries to solve a symbolic equation by applying applicable inverse functions or using polynomial roots. Pretty limited.

Example

julia> @symbolic x p
(x, p)

julia> eqn = exp(x) ~ 1//2
exp(x) ~ 1//2

julia> solve(eqn)  #
x ~ log(1//2)

julia> eqn = sin(x) + 2 ~ 3/2
sin(x) + 2 ~ 1.5

julia> solve(eqn)
x ~ -0.5235987755982989

julia> t1,v1,v2 = p[1], p[2], p[3]
(p[1], p[2], p[3])

julia> solve(sin(t1)/v1 ~ sin(x)/v2)
x ~ asin((sin(p[1]) / p[2]) * p[3])

Some expressions hold symbolic numbers, such as the first example. The numeric value of the resulting equation u can be found through u.rhs().

Polynomial expressions of degree 2 or more are handled differently, with return values. The quadratic and cubic formulas allow parameters to be used for 2nd and 3rd degree polynomials.

MTH229Lite.PlotlyLightLite.implicit_plot — Method

implicit_plot::SymbolicEquation

For equation written as F(x,y(x)) = 0 plot implicitly defined y.

Example

We simply use the symbolic parameter for the second variable

@symbolic u v
eq = u * v ~ u^3 + u^2 + u + 1

implicit_plot(eq)

To paramemterize a plot takes a bit of work to get the (x,y) values passed to the symbolic expression as a container in the first position:

@symbolic x p
u, v = x[1], x[2]
c,d,e,h = (p[i] for i in 1:4)
eq = u * v ~ c*u^3 + d*u^2 + e*u + h

implicit_plot((x,y) -> eq(tuple(x,y), (1,1,1,1)))

MTH229Lite.PlotlyLightLite.plot — Method

plot(eqn::SymbolicEquation, a::Real, b::Real; kwargs...)
plot(eqn::SymbolicEquation, ab::Interval; kwargs...)

Plot equation by plotting both left-hand side and right-hand side over the specified interval. Arguments label, linecolor and linewidth may specify two distinct values.

Example

# cf. https://www.chebfun.org/examples/roots/Tiger.html
@symbolic x
u = 2*exp(x/2) * (sin(5*x) + sin(101*x))
v = u - round(u)
ab = -2..1
p = plot(u ~ 0, ab; linecolor=("orange","blue"), linewidth=(nothing, 2), legend=false)
xs = solve(v ~ 0, ab)
filter!(x -> abs(v(x)) <= 1/8, xs) # no jumps, 345 left
scatter!(xs, u.(xs); markercolor="black")

MTH229Lite.bisection — Method

bisection(f, a, b) naive bisection algorithm with primitive graphic. Use find_zero(f, (a,b)) (or fzero(f, a, b)) for a more robust method.

MTH229Lite.fisheye — Method

fisheye(f) changes domain of function f to (-pi/2, pi/2)

MTH229Lite.lim — Method

lim(f, c; [dir="-"]) makes a table of values of f as x approaches c.

MTH229Lite.newton — Method

newton(f, x) easy to use Newton's method; derivative computed using f'

MTH229Lite.plotif — Method

plotif(f, g, a, b)

Plot f colored depending on g > 0 or g < 0.

MTH229Lite.rangeclamp — Function

rangeclamp(f, [hi], [lo]; replacement) returns f function which has the value of replacement when lo <= f(x) <= hi doesn't hold.

MTH229Lite.riemann — Method

riemann(f, a, b, n; method="right")

Simple Riemann sum, Method is one of "right", "left", "trapezoid", or "simpsons".

MTH229Lite.secant — Method

secant(f,a,b) returns a function computing the secant line of f between a and b

MTH229Lite.sign_chart — Method

signchart(f, a, b) identifies zeros of f (or discontinuities which jump over 0 and classifies sign change at each

MTH229Lite.tangent — Method

tangent(f,c) returns a function computing the tangent line of f at c

MTH229Lite.PlotlyLightLite — Module

PlotlyLightLite is a lightweight interface to the underlying PlotlyLight package (Fastest time-to-first-plot in Julia!), itself an "ultra-lightweight interface" to the Plotly javascript libraries. PlotlyLightLite supports some of the Plots.jl interface for plotting. This package may be of use as an alternative to Plots in resource-constrained environments, such as binder.

Plots.jl uses

positional arguments for data

keyword arguments for attributes

plot() is a workhorse with seriestype indicating which plot; there are also special methods (e.g. scatter(x,y) becomes plot(x,y; seriestype="scatter"))

PlotlyLight.jl uses

data::Vector{Config} to hold tracts of data for plotting

layout::Config to adjust layout

config::Config to adjust global configurations

Config very flexibly creates the underlying Javascript objects the plotly interface expects

Plot() is a workhorse with type acting like seriestype and also mode

PlotlyLightLite.jl has this dispatch for plot:

merge layout; merge config; pass kwargs to Config push onto data or merge onto last tract:

plot(; layout::Config?, config::Config, kwargs...)
plot!([p::Plot]; layout::Config?, config::Config?, kwargs...)

Line plot. connecting x,y (and possibly z). For 2D, use !isfinite values in y to break.

plot(x,y,[z]; kwargs...)
plot!([p::Plot], x, y, [z]; kwargs...

Data can be generated from a function:

plot(f::Function, ab; kwargs...) => plot(unzip(f, ab)...; kwargs...)
plot(f::Function, a, b; kwargs...) => plot(unzip(f, a, b)...; kwargs...)
plot!([p::Plot], f, ab, [b])

plot each function as lineplot:

plot(fs::Vector{Function}, a, [b]; kwargs...)
plot!([p::Plot], fs::Vector{Function}, a, [b]; kwargs...)

Note

Currently x, y make vectors; should matrices be supported using column vectors? "In Plots.jl, every column is a series, a set of related points which form lines, surfaces, or other plotting primitives. "

Parametric line plots, 2 or 3d

plot(fs::NTuple(N,Function), a, [b]; kwargs...)
plot!([p::Plot], fs::NTuple(N,Function), a, [b]; kwargs...)

Alternatively

plot(u::Function, v::Function, [w::Function], a, [b]; kwargs...)

In addition there are these plot constructors for higher-dimensional plots

contour

implicit_plot

heatmap

surface

wireframe

There are also numerous functions to modify attributes of an existing plot.

MTH229Lite.PlotlyLightLite.annotate! — Method

annotate!([p::Plot], x, y, txt; [color], [family], [pointsize], [halign], [valign])
annotate!([p::Plot], anns::Tuple;  kwargs...)

Add annotations to plot.

x, y, txt: text to add at (x,y)

color: text color

family: font family

pointsize: text size

halign: one of "top", "bottom"

valign: one of "left", "right"

rotation: angle to rotate

The x, y, txt values can be specified as 3 iterables or tuple of tuples.

MTH229Lite.PlotlyLightLite.arrow — Method

arrow(tails, vs; kwargs...)

Draw vectors vs anchored at tails. Hacked in support for 3D using combination of lines + cones.

Use Plotly attributes arrowcolor, arrowwidth,

MTH229Lite.PlotlyLightLite.circ3d — Method

circ3d(q, r, n̄; kwargs...)
circ3d!([p::Plot], q, r, n̄; kwargs...)

Plot circle in 3 dimensions with center at q, radius r, and perpendicular to normal vector n̄.

Example

q, n = [0,0,0], [0,0,1]
circ3d(q, 3, n)
arrow!(q, n)

Or a more complicated one:

Z(r, θ) = 4 - r
X(r, θ) = r * cos(θ)
Y(r, θ) = r * sin(θ)
rs = range(0,4, length=10)
θs = range(0, 2pi, length=100)
surface(X.(rs', θs), Y.(rs', θs), Z.(rs', θs); opacity=0.25)
q = [0,0,2]
n = [0,0,1]
r = 2
circ3d!(q, r, n; color="black", opacity=0.75)
q, v̂, ŵ = [0,0,0], [0,4,0], [0,0,4]
parallelogram!(q, v̂, ŵ; opacity=0.5, color=:yellow)

MTH229Lite.PlotlyLightLite.circle! — Method

circle([p::Plot], x0, x1, y0, y1; kwargs...)

Draw circle shape bounded in [x0, x1] × [y0,y1]. (Will adjust to non-equal sized boundary.)

Example

Use named tuple for line for boundary.

circle!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

MTH229Lite.PlotlyLightLite.contour — Method

contour(x, y, z; kwargs...)
contour!([p::Plot], x, y, z; kwargs...)
contour(x, y, f::Function; kwargs...)
contour!(x, y, f::Function; kwargs...)

Create contour function of f

MTH229Lite.PlotlyLightLite.current — Method

current()

Get current figure. A Plot object of PlotlyLight; UndefRefError if none.

Not typically needed, as it is implicit in most mutating calls, though may be convenient if those happen within a loop.

MTH229Lite.PlotlyLightLite.grid_layout — Method

grid_layout(ps::Array{<:Plot})

Layout an array of plots into a grid. Vectors become rows of plots.

Use Plot() to create an empty plot for a given cell.

Example

using DataFrames
n = 25
d = DataFrame(x=sin.(rand(n)), y=rand(n).^2, z = rand(n)) # assume all numeric
nms = names(d)
m = Matrix{Plot}(undef, length(nms), length(nms))

for i ∈ eachindex(nms)
    for j ∈ eachindex(nms)
        if j > i
            p = Plot()
        elseif j == i
            x = d[:,j]
            p = plot(x, type="histogram")
            xlabel!(p, nms[j])  # <<- not working in grid! as differe xaxis purposes
        else
            x = d[:,i]; y = d[:,j]
            p = scatter(x, y)
            xlabel!(p, nms[i]); ylabel!(p, nms[j])
        end
        m[i,j] = p
    end
end
grid_layout(m)

MTH229Lite.PlotlyLightLite.heatmap — Method

heatmap(x, y, z; kwargs...)
heatmap!([p::Plot], x, y, z; kwargs...)
heatmap(x, y, f::Function; kwargs...)
heatmap!(x, y, f::Function; kwargs...)

Create heatmap function of f

MTH229Lite.PlotlyLightLite.implicit_plot — Method

implicit_plot(f; xlims=(-5,5), ylims=(-5,5), legend=false, linewidth=2, kwargs...)
implicit_plot!([p::Plot], f; kwargs...)

For f(x,y) = ... plot implicitly defined y(x) from f(x,y(x)) = 0 over range specified by xlims and ylims.

Example

f(x,y) = x * y - (x^3 + x^2 + x + 1)
implicit_plot(f)

(Basically just contour plot with levels=0 and points detremined by extrema of xlims and ylims.)

MTH229Lite.PlotlyLightLite.legend! — Function

legend!([p::Plot], legend::Bool) hide/show legend

MTH229Lite.PlotlyLightLite.parallelogram — Method

parallelogram(q, v̄, w̄; kwargs...)
parallelogram!([p::Plot], q, v̄, w̄; kwargs...)

Plot parallelogram formed by two vectors, v̄ and w̄, both anchored at point q.

MTH229Lite.PlotlyLightLite.plot! — Method

plot!([p::Plot], x, y; kwargs...)
plot!([p::Plot], f, a, [b]; kwargs...)
plot!([p::Plot], f; kwargs...)

Used to add a new tract to an existing plot. Like Plots.plot!. See plot for argument details.

MTH229Lite.PlotlyLightLite.plot — Method

plot(x, y, [z]; [linecolor], [linewidth], [legend], kwargs...)
plot(f::Function, a, [b]; kwargs...)

Create a line plot.

Returns a Plot instance from PlotlyLight

x,y points to plot. NaN values in y break the line

a, b: the interval to plot a function over can be given by two numbers or if just a then by extrema(a).

linecolor: color of line

linewidth: width of line

label in legend

Other keyword arguments include width and height, xlims and ylims, legend, aspect_ratio.

Provides an interface like Plots.plot for plotting a function f using PlotlyLight. This just scratches the surface, but PlotlyLight allows full access to the underlying JavaScript library library.

The provided "Plots-like" functions are plot, plot!, scatter!, scatter, annotate!, title!, xlims! and ylims!.

Example

p = plot(sin, 0, 2pi; legend=false)
plot!(cos)
# add points
x0 = [pi/4, 5pi/4]
scatter!(x0, sin.(x0), markersize=10)
# add text
annotate!(tuple(zip(x0, sin.(x0), ("A", "B"))...), halign="left", pointsize=12)
title!("Sine and cosine and where they intersect in [0,2π]")
# adjust limits
ylims!((-3/2, 3/2))
# add shape
y0, y1 = extrema(p).y
[rect!(xᵢ-0.1, xᵢ+0.1, y0, y1, fillcolor="gray", opacity=0.2) for xᵢ ∈ x0]
# display plot
p

Warning

You may need to run the first plot cell twice to see an image.

MTH229Lite.PlotlyLightLite.plot — Method

plot(; layout::Config?, config::Config?, kwargs...)

Pass keyword arguments through Config and onto PlotlyLight.Plot.

MTH229Lite.PlotlyLightLite.plot — Method

plot((f,g), a, b; kwargs...)
plot!([p::PLot], (f,g), a, b; kwargs...)

Make parametric plot from tuple of functions, f and g.

MTH229Lite.PlotlyLightLite.quiver! — Function

quiver!([p::Plot], x, y, txt=nothing; quiver=(dx, dy), kwargs...)
quiver(x, y, txt=nothing; quiver=(dx, dy), kwargs...)

Draw 2d arrows. See arrow! for a single arrow.

(x,y) are tail positions, optionally labeled by txt

quiver specifies vector part of arrow

kwargs processarrowhead::Int?,arrowwidth::Int?,arrowcolor`

Example

ts = range(0, 2pi, length=100)
p = plot(sin.(ts), cos.(ts), linecolor="red")
ts = range(0, 2pi, length=10)
quiver!(p, cos.(ts), sin.(ts), quiver=(-sin.(ts), cos.(ts)), arrowcolor="red")
p

This example shows how text can be rotated with angles in degrees and positive angles measured in a clockwise direction.

ts = range(0, 2pi, 100)
p = plot(cos.(ts), sin.(ts), linecolor="red", aspect_ratio=:equal,
    linewidth=20, opacity=0.2)

txt = split("The quick brown fox jumped over the lazy dog")
ts = range(0, 360, length(txt)+1)[2:end]
for (i,t) ∈ enumerate(reverse(ts))
    quiver!(p, [cosd(t)],[sind(t)],txt[i],
            quiver=([0],[0]),
            textangle=90-t,
            font=(size=20,))
end
xaxis!(zeroline=false); yaxis!(zeroline=false) # remove zerolines
p

3d arrows

3d arrows are possible using arrows!.

MTH229Lite.PlotlyLightLite.rect! — Method

rect!([p::Plot], x0, x1, y0, y1; kwargs...)

Draw rectangle shape.

Example

Use named tuple for line for boundary.

rect!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

MTH229Lite.PlotlyLightLite.scatter! — Method

scatter(x, y, [z]; [markershape], [markercolor], [markersize], kwargs...)
scatter!([p::Plot], x, y, [z]; kwargs...)

Place point on a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red"

markersize: size, as an integer

MTH229Lite.PlotlyLightLite.scatter — Method

scatter(x, y; kwargs...) see scatter!

MTH229Lite.PlotlyLightLite.scroll_zoom! — Method

scrollzoom!([p], x::Bool) turn on/off scrolling to zoom

MTH229Lite.PlotlyLightLite.size! — Method

size!([p::Plot]; [width], [height]) specify size of plot figure

MTH229Lite.PlotlyLightLite.skirt — Method

skirt(q, v, f::Function; kwargs...)
skirt!([p::Plot], q, v, f::Function; kwargs...)
skirt!([p::Plot], xs, ys, zs, f::Function; kwargs...)

Along a path (xs, ys, zs) plot a skirt between the path and (xs, ys, f(xs, ys)). The case of a path described by a vector, v, anchored at a point q has a special method.

Example

x(r, θ) = r*cos(θ)
y(r, θ) = r*sin(θ)
f(x,y) = 4 - x^2 - y^2
rs, θs = range(0,2,length=10), range(0, 2pi, length=20)
xs, ys = x.(rs', θs), y.(rs', θs)
zs = f.(xs, ys)
surface(xs, ys, zs, opacity=.2, showscale=false)

q = [0,0,0]
v = [2,0,0]
d = skirt!(q, v, f; opacity=0.6)

t = range(0, 1, length=100)
xs = 2 * t .* sin.(t*pi/2)
ys = 2 * t .* cos.(t*pi/2)
zs = zero.(xs)
skirt!(xs, ys, zs, f; color="blue", opacity=0.6)
skirt!(xs, -ys, zs, f; color="blue", opacity=0.6)

MTH229Lite.PlotlyLightLite.surface — Method

surface(x, y, z; kwargs...)
surface!(x, y, z; kwargs...)
surface(x, y, f::Function; kwargs...)
surface!(x, y, f::Function; kwargs...)

Create surface plot. Pass zcontour=true to add contour plot projected onto the z axis.

Example

f(x,y) = 4 - x^2 - y^2
xs = ys = range(-2, 2, length=50)
surface(xs, ys, f)  # zs = f.(xs', ys)

Extended help

From https://discourse.julialang.org/t/3d-surfaces-time-slider/109673

z1 = Vector[[8.83, 8.89, 8.81, 8.87, 8.9, 8.87],
                       [8.89, 8.94, 8.85, 8.94, 8.96, 8.92],
                       [8.84, 8.9, 8.82, 8.92, 8.93, 8.91],
                       [8.79, 8.85, 8.79, 8.9, 8.94, 8.92],
                       [8.79, 8.88, 8.81, 8.9, 8.95, 8.92],
                       [8.8, 8.82, 8.78, 8.91, 8.94, 8.92],
                       [8.75, 8.78, 8.77, 8.91, 8.95, 8.92],
                       [8.8, 8.8, 8.77, 8.91, 8.95, 8.94],
                       [8.74, 8.81, 8.76, 8.93, 8.98, 8.99],
                       [8.89, 8.99, 8.92, 9.1, 9.13, 9.11],
                       [8.97, 8.97, 8.91, 9.09, 9.11, 9.11],
                       [9.04, 9.08, 9.05, 9.25, 9.28, 9.27],
                       [9, 9.01, 9, 9.2, 9.23, 9.2],
                       [8.99, 8.99, 8.98, 9.18, 9.2, 9.19],
                       [8.93, 8.97, 8.97, 9.18, 9.2, 9.18]]
xs , ys = 1:length(z1[1]), 1:length(z1) # needed here given interface chosen
surface(xs, ys, z1, colorscale="Viridis")
surface!(xs, ys, map(x -> x .+ 1, z1), colorscale="Viridis", showscale=false, opacity=0.9)
surface!(xs, ys, map(x -> x .- 1, z1), colorscale="Viridis", showscale=false, opacity=0.9)

Julia users would typically use a matrix to hold the z data, but Javascript users would expect a vector of vectors, as above. As PlotlyLight just passes on the data to Javascript, the above is perfectly acceptable. Indeed PlotlyLightLite converts matrices to this format in such plots.

(The keyword arguments above come from Plotly, not Plots.)

A parameterized surface can be displayed. Below the unzip function returns 3 matrices specifying the surface described by the vector-valued function r.

r1, r2 = 2, 1/2
r(u,v) = ((r1 + r2*cos(v))*cos(u), (r1 + r2*cos(v))*sin(u), r2*sin(v))
us = vs = range(0, 2pi, length=25)
xs, ys, zs = unzip(us, vs, r)

surface(xs, ys, zs)

Plotting a vertical plane poses a slight challenge, as we can't parameterize as z=f(x,y). Here we intersect a surface with the plane ax + by + 0z = d and add a trace for the intersection of the two surfaces.

f(x, y) = 4 - (x^2 + y^2)

# surface
xs =  ys = range(-2, 2, length=100)
zs = f.(xs', ys)

# (vertical) plane
m, M = extrema(zs)
zzs = range(m, M, length=2)
Xs = ((x,z) -> x).(xs', zzs)
Zs = ((x,z) -> z).(xs', zzs)
a, b, c, d = 1, 1, 0, 1
plane(x,z) = (d - a*x - c*z) / b
Ys = plane.(xs', zzs)

# intersection
g(t) = (d - a*t) / b
γ(t) = (t, y(t), f(t, y(t)))

surface(xs, ys, zs)
surface!(Xs, Ys, Zs, opacity=0.2)
plot!(unzip(γ.(xs))...; linewidth=3)

MTH229Lite.PlotlyLightLite.title! — Method

title!([p::Plot], txt)
xlabel!([p::Plot], txt)
ylable!([p::Plot], txt)
zlabel!([p::Plot], txt)

Set plot title.

MTH229Lite.PlotlyLightLite.unzip — Method

unzip(v, [vs...])
unzip(f::Function, a, b)
unzip(a, b, F::Function)

Reshape data to x,y,[z] mode.

In its basic use, zip takes two vectors, pairs them off, and returns an iterator of tuples for each pair. For unzip a vector of same-length vectors is "unzipped" to return two (or more) vectors.

The function version applies f to a range of points over (a,b) and then calls unzip. This uses the adapted_grid function from PlotUtils.

The function version with F computes F(a', b) and then unzips. This is used with parameterized surface plots

This uses the invert function of SplitApplyCombine.

MTH229Lite.PlotlyLightLite.wireframe — Method

wireframe(x, y, z; kwargs...)
wireframe(x, y, f::Function; kwargs...)
wireframe!([p::Plot], x, y, z; kwargs...)
wireframe!([p::Plot], x, y, f::Function; kwargs...)

Create wireframe.

Example

f(x, y) = 4 - x^2 - y^2
xs = ys = range(-2, 2, length=100)
surface(xs, ys, f)
wireframe!(xs, ys, f)

MTH229Lite.PlotlyLightLite.xaxis! — Method

xaxis!([p::Plot]; kwargs...)
yaxis!([p::Plot]; kwargs...)
zaxis!([p::Plot]; kwargs...)

Adjust ticks on chart.

ticks: a container or range

ticklabels: optional labels (same length as ticks)

showticklabels::Bool

MTH229Lite.PlotlyLightLite.xlims! — Method

xlims!(p, lims) set x limits of plot

MTH229Lite.PlotlyLightLite.ylims! — Method

ylims!(p, lims) set y limits of plot

MTH229Lite.PlotlyLightLite.ziptie — Method

ziptie(xs, ys, zs, xs′, ys′, zs′; kwargs...)
ziptie!([p::Plot], xs, ys, zs, xs′, ys′, zs′; kwargs...)

surface created by connecting points along two paths given by (xs, ys, zs) and (xs′, ys′, zs′). All vectors must be same length. Mesh is created by zipping together points on the two curves.

Example

r(t) = (sin(t), cos(t), t)
s(t) = (sin(t+pi), cos(t+pi), t)
ts = range(0, 4pi, length=100)
ziptie(unzip(r.(ts))..., unzip(s.(ts))...;
       color="green", opacity=.25, showscale=false)

MTH229Lite.PlotlyLightLite.zlims! — Method

zlims!(p, lims) set z limits of plot

MTH229Lite.PlotlyLightLite.★ — Method

★(q, xs, ys, zs; kwargs...)
★!([p::Plot], q, xs, ys, zs; kwargs...)

A star connected region has an origin, q, for which each boundary point (described by (xs, ys, zs) is accessible by a ray for q which does not cross the boundary.

Example

pts = 5
Δ = 2pi/pts/2
a, A = 1, 3
q = [0,0,0]
ts = range(0, 2pi, length=pts+1)
ps = [(A*[cos(t),sin(t),0], a*[cos(t+Δ), sin(t+Δ), 0]) for t in ts]
xs, ys, zs = unzip(collect(Base.Iterators.flatten(ps)))
★(q, xs, ys, zs)