MTH229Lite

Documentation for MTH229Lite.

MTH229Lite.MTH229Lite

MTH229Lite.annotate!

MTH229Lite.bisection

MTH229Lite.fisheye

MTH229Lite.lim

MTH229Lite.newton

MTH229Lite.plot

MTH229Lite.plot!

MTH229Lite.plotif

MTH229Lite.rangeclamp

MTH229Lite.riemann

MTH229Lite.scatter

MTH229Lite.secant

MTH229Lite.sign_chart

MTH229Lite.tangent

MTH229Lite.title!

MTH229Lite.xlims!

MTH229Lite.ylims!

MTH229Lite.MTH229Lite — Module

MTH229Lite

Support files for MTH229. Unlike the MTH229 package, this omits SymPy and has a simplified plotting setup for running within resource constrained environments, such as binder.org.

Exports:

e aliased to exp(1)

tangent(f,c) and secant(f,a,b) for creating tangent and secant line functions

rangeclamp to restrict large y values; fisheye to change x axis scale

lim to explore limits numerically

sign_chart to explore sign changes numerically

riemann a simple Riemann sum function. This package also exports QuadGK.

bisection and newton for solving f(x) = 0. This package also exports Roots which exposes find_zero and find_zeros (and their aliasesfzero, fzeros).

The operator ' is overloaded for functions to find the derivative using the FowardDiff package

The SimpleExpressions package is included to create simple expressions to use in place of functions. There is no SymPy package provided, as is done with MTH229.

The plot and plot! functions provided are modeled after those in Plots.jl but utilize PlotlyLight as a plotting backend, as it requires far fewer resources than Plots. See the docstring for plot for an example.

MTH229Lite.annotate! — Method

annotate!([p::Plot], x, y, txt; size, color, textposition, kwargs...)
annotate!([p::Plot], anns::Tuple; size, color, textposition, kwargs...)

Add annotations to plot

size: text size

color: text color

textposition: one of "top", "bottom", "left", "right", or combinations with _

MTH229Lite.bisection — Method

bisection(f, a, b) naive bisection algorithm with primitive graphic. Use find_zero(f, (a,b)) (or fzero(f, a, b)) for a more robust method.

MTH229Lite.fisheye — Method

fisheye(f) changes domain of function f to (-pi/2, pi/2)

MTH229Lite.lim — Method

lim(f, c; [dir="-"]) makes a table of values of f as x approaches c.

MTH229Lite.newton — Method

newton(f, x) easy to use Newton's method; derivative computed using f'

MTH229Lite.plot! — Method

plot!([p::Plot], f; kwargs...)
plot!([p::Plot], x, y; kwargs...)

Used to add a new tract to an existing plot. Like Plots.plot!

MTH229Lite.plot — Method

plot(x, y; kwargs...)
plot(f::Function, a, b; kwargs...)

Create a plot. Returns a Plot instance from PlotlyLight

Keyword arguments include width and height, xlims and ylims, legend.

Provides an interface like Plots.plot for plotting a function f using PlotlyLight. This just scratches the surface, but PlotLight allows full acces to the underlying JavaScript library.

The provided "Plots" like functions are plot, plot!, scatter, scatter!, annotate!, title!, xlims! and ylims!.

Example

plot(sin, 0, 2pi)
plot!(cos)
x0 = [pi/4, 5pi/4]
scatter!(x0, sin.(x0), markersize=10)
title!("Sine and cosine and where they intersect in [0,2π]")

MTH229Lite.plotif — Method

plotif(f, g, a, b)

Plot f colored depending on g >= 0 or not.

MTH229Lite.rangeclamp — Function

rangeclam(f, [hi], [lo]; replacement) returns f function which has the value of replacement when lo <= f(x) <= hi doesn't hold.

MTH229Lite.riemann — Method

riemann(f, a, b, n; method="right")

Simple Riemann sum, Method is one of "right", "left", "trapezoid", or "simpsons".

MTH229Lite.scatter — Method

scatter(x, y; markershape, markercolor, markersize, kwargs...)
scatter!([p::Plot], x, y; kwargs...)

Place point on a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red"

markersize: size, as an integer

MTH229Lite.secant — Method

secant(f,a,b) returns a function computing the secant line of f between a and b

MTH229Lite.sign_chart — Method

signchart(f, a, b) identifies zeros of f (or discontinuities which jump over 0 and classifies sign change at each

MTH229Lite.tangent — Method

tangent(f,c) returns a function computing the tangent line of f at c

MTH229Lite.title! — Method

title!([p::Plot], txt)

Set plot title.

MTH229Lite.xlims! — Method

xlims!(p, lims) set x limits of plot

MTH229Lite.ylims! — Method

ylims!(p, lims) set y limits of plot