MTH229Lite.MTH229Lite — Module

MTH229Lite

Support files for MTH229. Unlike the MTH229 package, this package omits SymPy and has a simplified plotting setup for running within resource-constrained environments, such as binder.org (e.g. mth229lite).

Exports:

e aliased to exp(1)

tangent(f,c) and secant(f,a,b) for creating tangent and secant line functions

rangeclamp to restrict large y values; fisheye to change x axis scale; unzip to manipulate plotting limits

lim to explore limits numerically

sign_chart to explore sign changes numerically

bisection and newton for solving f(x) = 0. This package also exports Roots which exposes find_zero and find_zeros (and their aliasesfzero, fzeros).

riemann a simple Riemann sum function. This package also exports QuadGK for real-world usage.

The operator ' is overloaded for functions to find the derivative using the ForwardDiff package. See D also.

The SimpleExpressions package is included to create simple expressions to use in place of functions. There is no SymPy package provided, as is done with MTH229.

The plot and plot! functions provided are modeled after those in Plots.jl but utilize PlotlyLight as a plotting backend, as it requires far fewer resources than Plots. See the docstring for plot for an example.

MTH229Lite.Interval — Type

Interval(a,b)

A means to specify an interval in the form a..b

There are methods for

plot(eq, I) (two plots)

solve(eq, I) (dispatch to find_zeros)

sign_chart, riemann, quadgk

There are parsing gotchas

-1..2 is okay

-2..-1 is not, try -2..(-1)

-1..x^2 will parse as Interval(-1,x^2), also -1..x+2 becomes Interval(-1, x+2), as .. has low precedence.

Displays with unicode brackets to disambiguate from vector or tuple:

julia> -2..1
⟦-2, 1⟧

CommonSolve.solve — Method

solve(ex::SymbolicEquation, x₀, args...; kwargs...)
solve(ex::SymbolicEquation, x₀::Interval; kwargs...)`

Numerically solve an equation specified with a @symbolic value.

The find_zero or find_zeros function is used, the latter when x₀ is of type Interval (specified with the .. infix operator.

Example

julia> solve(e^x ~ x^4, 10)       # solution near 10
8.6131694564414

julia> solve(e^x ~ x^4, -10, 10)  # some solution in bracketing interval
-0.8155534188089607

julia> solve(e^x ~ x^4, -10..10)  # all solutions within [-10, 10]
3-element Vector{Float64}:
 -0.8155534188089606
  1.4296118247255554
  8.6131694564414

CommonSolve.solve — Method

solve(ex::SymbolicEquation)

Tries to solve a symbolic equation by applying applicable inverse functions or using polynomial roots. Pretty limited.

Example

julia> @symbolic x p
(x, p)

julia> eqn = exp(x) ~ 1//2
exp(x) ~ 1//2

julia> solve(eqn)  #
x ~ log(1//2)

julia> eqn = sin(x) + 2 ~ 3/2
sin(x) + 2 ~ 1.5

julia> solve(eqn)
x ~ -0.5235987755982989

julia> t1,v1,v2 = p[1], p[2], p[3]
(p[1], p[2], p[3])

julia> solve(sin(t1)/v1 ~ sin(x)/v2)
x ~ asin((sin(p[1]) / p[2]) * p[3])

Some expressions hold symbolic numbers, such as the first example. The numeric value of the resulting equation u can be found through u.rhs().

Polynomial expressions of degree 2 or more are handled differently, with return values. The quadratic and cubic formulas allow parameters to be used for 2nd and 3rd degree polynomials.

MTH229Lite.annotate! — Method

annotate!([p::Plot], x, y, txt; [color], [family], [pointsize], [halign], [valign])
annotate!([p::Plot], anns::Tuple;  kwargs...)

Add annotations to plot.

x, y, txt: text to add at (x,y)

color: text color

family: font family

pointsize: text size

halign: one of "top", "bottom"

valign: one of "left", "right"

rotation: angle to rotate

The x, y, txt values can be specified as 3 iterables or tuple of tuples.

MTH229Lite.bisection — Method

bisection(f, a, b) naive bisection algorithm with primitive graphic. Use find_zero(f, (a,b)) (or fzero(f, a, b)) for a more robust method.

MTH229Lite.circle! — Method

circle([p::Plot], x0, x1, y0, y1; kwargs...)

Draw circle shape bounded in [x0, x1] × [y0,y1]. (Will adjust to non-equal sized boundary.)

Example

Use named tuple for line for boundary.

circle!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

MTH229Lite.fisheye — Method

fisheye(f) changes domain of function f to (-pi/2, pi/2)

MTH229Lite.gcf — Method

gcf()

Get current figure. A Plot object of PlotlyLight. Not typically needed, as it is implicit in most mutating calls, though may be convenient if those happen within a loop.

MTH229Lite.grid_layout — Method

grid_layout(ps::Array{<:Plot})

Layout an array of plots into a grid. Vectors become rows of plots.

Use Plot() to create an empty plot for a given cell.

Example

using DataFrames
n = 25
d = DataFrame(x=sin.(rand(n)), y=rand(n).^2, z = rand(n)) # assume all numeric
nms = names(d)
m = Matrix{Plot}(undef, length(nms), length(nms))

for i ∈ eachindex(nms)
    for j ∈ eachindex(nms)
        if j > i
            p = Plot()
        elseif j == i
            x = d[:,j]
            p = plot(x, type="histogram")
            xlabel!(p, nms[j])  # <<- not working in grid! as differe xaxis purposes
        else
            x = d[:,i]; y = d[:,j]
            p = scatter(x, y)
            xlabel!(p, nms[i]); ylabel!(p, nms[j])
        end
        m[i,j] = p
    end
end
grid_layout(m)

MTH229Lite.legend! — Function

legend!([p::Plot], legend::Bool) hide/show legend

MTH229Lite.lim — Method

lim(f, c; [dir="-"]) makes a table of values of f as x approaches c.

MTH229Lite.newton — Method

newton(f, x) easy to use Newton's method; derivative computed using f'

MTH229Lite.plot! — Method

plot!([p::Plot], x, y; kwargs...)
plot!([p::Plot], f; kwargs...)

Used to add a new tract to an existing plot. Like Plots.plot!

MTH229Lite.plot — Method

plot(x, y, [z]; [linecolor], [linewidth], [legend], kwargs...)
plot(f::Function, a, b; kwargs...)

Create a line plot.

Returns a Plot instance from PlotlyLight

x,y points to plot. NaN values in y break the line

linecolor: color of line

linewidth: width of line

label

Other keyword arguments include width and height, xlims and ylims, legend, aspect_ratio.

Provides an interface like Plots.plot for plotting a function f using PlotlyLight. This just scratches the surface, but PlotlyLight allows full access to the underlying JavaScript library library.

The provided "Plots-like" functions are plot, plot!, scatter!, scatter, annotate!, title!, xlims! and ylims!.

Example

p = plot(sin, 0, 2pi; legend=false)
plot!(cos)
# add points
x0 = [pi/4, 5pi/4]
scatter!(x0, sin.(x0), markersize=10)
# add text
annotate!(tuple(zip(x0, sin.(x0), ("A", "B"))...), halign="left", pointsize=12)
title!("Sine and cosine and where they intersect in [0,2π]")
# adjust limits
ylims!((-3/2, 3/2))
# add shape
y0, y1 = extrema(p).y
[rect!(xᵢ-0.1, xᵢ+0.1, y0, y1, fillcolor="gray", opacity=0.2) for xᵢ ∈ x0]
# display plot
p

Warning

You may need to run the first plot cell twice to see an image.

MTH229Lite.plot — Method

plot(eqn::SymbolicEquation, a::Real, b::Real; kwargs...)
plot(eqn::SymbolicEquation, ab::Interval; kwargs...)

Plot equation by plotting both left-hand side and right-hand side over the specified interval. Argument linecolor and linewidth may specify two distinct values.

Example

# cf. https://www.chebfun.org/examples/roots/Tiger.html
@symbolic x
u = 2*exp(x/2) * (sin(5*x) + sin(101*x))
v = u - round(u)
ab = -2..1
p = plot(u ~ 0, ab; linecolor=("orange","blue"), linewidth=(nothing, 2), legend=false)
xs = solve(v ~ 0, ab)
filter!(x -> abs(v(x)) <= 1/8, xs) # no jumps, 345 left
scatter!(xs, u.(xs); markercolor="black")

MTH229Lite.plotif — Method

plotif(f, g, a, b)

Plot f colored depending on g > 0 or g < 0.

MTH229Lite.quiver! — Function

quiver!([p::Plot], x, y, txt=nothing; quiver=(dx, dy), kwargs...)
quiver(x, y, txt=nothing; quiver=(dx, dy), kwargs...)

(x,y) are tail positions, optionally labeled by txt

quiver specifies vector part of arrow

kwargs processarrowhead::Int?,arrowwidth::Int?,arrowcolor`

Example

ts = range(0, 2pi, length=100)
p = plot(sin.(ts), cos.(ts), linecolor="red")
ts = range(0, 2pi, length=10)
quiver!(p, cos.(ts), sin.(ts), quiver=(-sin.(ts), cos.(ts)), arrowcolor="red")
p

This example shows how text can be rotated with angles in degrees and positive angles measured in a clockwise direction.

ts = range(0, 2pi, 100)
p = plot(cos.(ts), sin.(ts), linecolor="red", aspect_ratio=:equal,
    linewidth=20, opacity=0.2)

txt = split("The quick brown fox jumped over the lazy dog")
ts = range(0, 360, length(txt)+1)[2:end]
for (i,t) ∈ enumerate(reverse(ts))
    quiver!(p, [cosd(t)],[sind(t)],txt[i],
            quiver=([0],[0]),
            textangle=90-t,
            font=(size=20,))
end
xaxis!(zeroline=false); yaxis!(zeroline=false) # remove zerolines
p

MTH229Lite.rangeclamp — Function

rangeclamp(f, [hi], [lo]; replacement) returns f function which has the value of replacement when lo <= f(x) <= hi doesn't hold.

MTH229Lite.rect! — Method

rect!([p::Plot], x0, x1, y0, y1; kwargs...)

Draw rectangle shape.

Example

Use named tuple for line for boundary.

rect!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

MTH229Lite.riemann — Method

riemann(f, a, b, n; method="right")

Simple Riemann sum, Method is one of "right", "left", "trapezoid", or "simpsons".

MTH229Lite.scatter! — Method

scatter(x, y, [z]; [markershape], [markercolor], [markersize], kwargs...)
scatter!([p::Plot], x, y, [z]; kwargs...)

Place point on a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red"

markersize: size, as an integer

MTH229Lite.scatter — Method

scatter(x, y; kwargs...) see scatter!

MTH229Lite.scroll_zoom! — Method

scrollzoom!([p], x::Bool) turn on/off scrolling to zoom

MTH229Lite.secant — Method

secant(f,a,b) returns a function computing the secant line of f between a and b

MTH229Lite.sign_chart — Method

signchart(f, a, b) identifies zeros of f (or discontinuities which jump over 0 and classifies sign change at each

MTH229Lite.size! — Method

size!([p::Plot]; [width], [height]) specify size of plot figure

MTH229Lite.tangent — Method

tangent(f,c) returns a function computing the tangent line of f at c

MTH229Lite.title! — Method

title!([p::Plot], txt)

Set plot title.

MTH229Lite.unzip — Method

unzip(v, [vs...])
unzip(f::Function, a, b)

Take container of points, return vector of corrodinated. Reverse of zip. Function version applies f to a range of points over (a,b) and then calls unzip.

MTH229Lite.xaxis! — Method

xticks!([p::Plot]; kwargs...)
yticks!([p::Plot]; kwargs...)

Adjust ticks on chart.

ticks: a container or range

ticklabels: optional labels (same length as ticks)

showticklabels::Bool

MTH229Lite.xlims! — Method

xlims!(p, lims) set x limits of plot

MTH229Lite.ylims! — Method

ylims!(p, lims) set y limits of plot